requirements.txt
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0
python-dotenv>=1.0.0
